server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-mysqldb==1.0.1
python-dotenv==1.0.0
bcrypt==4.0.1
Pillow==10.0.0
pyjwt==2.8.0
python-dateutil==2.8.2
Werkzeug==2.3.7